requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
dacite==1.8.0
elastic-transport==8.4.0
elasticsearch==8.7.0
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pluggy==1.0.0
pytest==7.3.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.9.1
